+++ noodles-fasta/src/repository/adapter.rs
@@ -3,7 +3,7 @@ use std::io;
 use crate::Record;
 
 /// A repository adapter.
-pub trait Adapter {
+pub trait Adapter: Send {
     /// Returns the record with the given name.
     fn get(&mut self, name: &[u8]) -> Option<io::Result<Record>>;
 }