+++ noodles-fasta/src/repository.rs
@@ -5,7 +5,11 @@ pub mod adapters;
 
 pub use self::adapter::Adapter;
 
-use std::{collections::HashMap, fmt, io, rc::Rc, sync::RwLock};
+use std::{
+    collections::HashMap,
+    fmt, io,
+    sync::{Arc, RwLock},
+};
 
 use super::record::Sequence;
 
@@ -15,7 +19,7 @@ struct AdapterCache {
 }
 
 /// A caching sequence repository.
-pub struct Repository(Rc<RwLock<AdapterCache>>);
+pub struct Repository(Arc<RwLock<AdapterCache>>);
 
 impl Repository {
     /// Creates a sequence repository.
@@ -23,7 +27,7 @@ impl Repository {
     where
         A: Adapter + 'static,
     {
-        Self(Rc::new(RwLock::new(AdapterCache {
+        Self(Arc::new(RwLock::new(AdapterCache {
             adapter: Box::new(adapter),
             cache: HashMap::new(),
         })))